Devjp

+++ database/migrations/2021_05_05_175623_alter_table_users_add_columns.php
@@ -14,14 +14,16 @@ class AlterTableUsersAddColumns extends Migration
     public function up()
     {
         Schema::enableForeignKeyConstraints();
+
         Schema::table('users', function (Blueprint $table) {
-            $table->dropForeign('fk_oficinas_usuarios');
-            $table->dropColumn('oficinas_id');
+            if (Schema::hasColumn('users', 'oficinas_id')) {
+                $table->dropColumn('oficinas_id');
+            }
             //
-            $table->integer('id_padre', 11)->unsigned()->nullable()->after('roles_id');
-            $table->integer('id_company', 11)->unsigned()->nullable()->after('roles_id');
-            $table->integer('privilegios_ami', 1)->nullable()->after('roles_id');
-            $table->integer('privilegios_hego', 1)->nullable()->after('roles_id');
+            $table->integer('hego')->nullable()->after('roles_id');
+            $table->integer('creausuarios')->nullable()->after('roles_id');
+            $table->integer('id_padre')->unsigned()->nullable()->after('roles_id');
+            $table->integer('id_company')->unsigned()->nullable()->after('roles_id');
             //
             $table->foreign('id_padre')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('id_company')->references('id')->on('users')->onDelete('cascade');
@@ -35,10 +37,23 @@ class AlterTableUsersAddColumns extends Migration
      */
     public function down()
     {
+        Schema::enableForeignKeyConstraints();
+        
         Schema::table('users', function (Blueprint $table) {
             $table->dropForeign('users_id_padre_foreign');
             $table->dropForeign('users_id_company_foreign');
-            $table->dropColumn(['id_padre', 'id_company', 'privilegios_ami', 'privilegios_hego']);
+            if (Schema::hasColumn('users', 'id_padre')) {
+                $table->dropColumn('id_padre');
+            }
+            if (Schema::hasColumn('users', 'id_company')) {
+                $table->dropColumn('id_company');
+            }
+            if (Schema::hasColumn('users', 'creausuarios')) {
+                $table->dropColumn('creausuarios');
+            }
+            if (Schema::hasColumn('users', 'hego')) {
+                $table->dropColumn('hego');
+            }
         });
     }
 }
